fix: Fix PiedrabrujaSpider price selector
The store has a similar problem that Flexogames, when a listing is on sale,
the .price element contains both the original price and the sale price, so
the price text was concatenating both numbers.

In addition, the HTML send is malformed, every <li> element has a <div>
that isn't closed, so every listing contains every following items content.
This prohibits the resuse of the same solution used in Flexogames.