refactor(EntrejuegosSpider): add #purchasable? method
While handling the case where items are listed without a price tag
I saw that the #in_stock? method was the wrong abstraction/name.

The #purchasable? method acts as an aggregator of all the predicate
methods that checks for an item availability or purchasability.

I also though that #purchasable? was more precise than #available?