fix(PrestaShop::Spider): handle items with fallback images
When stores upload a product without image, a fallback image tag is generated.
The fallback `img[src]` breaks the `#get_title`'s slug strategy.

This commits handles this case with:
- check for fallback images (alt presence)
- new method to get title from product url slug

I tried to replace `#title_from_slug`  for the nmew `#product_slug_title`,
but that broke the AldeaJuegosSpider discounted test.